perf(modules): update the academics, student, and blocknote module
academics, update the data validation, update the routes and add in the authentication middleware,
improve the logic in the controller and the swagger file configuration. Update the student
validations, authentication and add in the endpoint to get a student by id. Add the authentication
middleware in the blocknote to let authenticated users take notes in the system