Merge 2: code stability (#19)
* Create student validator, CRD operations

* feat(blocknote): blocknote

ni

* feat(created the blocknote module): students and teachers can add task

Task can be added and set deadline, students and teachers can also add notes

* feat(added deadline to the schema): deadline was added

Deadline can be added on tasks

* Leaves, TimeTable and Payroll tasks

* Student CRUD operation, removal of 'password' from Student model

* Add student routes

* HRM with validators and Swager doc

* Added student route, corrected Gender enum for student validator

* refactor(human resources): refactor the HR module and generate student's swagger file

refactor the HR components into a single module, generate the swagger file of the student module

* refactor(module): structure the academic

refactor the academics components into a single module, resolved merged conflics, add the swagger
file for the academic year

* fix(module): merge conflic, code structure

resolved the merge conflic, update the routes, and add the swagger configuration for the blocknote

* feat(validation): inputs are now validated

inputs are validated so that only the correct data is sent

* Add param to delete and update route

* feat(file): Incident module implement with crud ops (#16)

feat(file): added the crud operations  on the incident modules

* Updated student controller and route file

* testing

* perf(component): improve the database and the admin component

adjust the prisma schema accordingly to let the database beign consistant, improve the routes of the
incident and school event componenets by updating the routes and the swagger file

* fix(academics module): adjust the fields accordingly with respect to the database

adjust the fields by deleting the class model and maintaining the classes model because they where
refering the same model, avoiding keywords naming like that of "class" and adjust the controller
academics to match with the prisma schema

* perf(modules): update the academics, student, and blocknote module

academics, update the data validation, update the routes and add in the authentication middleware,
improve the logic in the controller and the swagger file configuration. Update the student
validations, authentication and add in the endpoint to get a student by id. Add the authentication
middleware in the blocknote to let authenticated users take notes in the system

* perf(module): the human resource code improved

improve the routes and put in the authentication of users, adjust the controller accordingly and
update the swagger file of the academic module

---------

Co-authored-by: Jane Ahone <jane.eloundou@gmail.com>
Co-authored-by: OBI OBI ETCHU JNR <obi.junior@icloud.com>
Co-authored-by: TIWA <ntiwa11@gmail.com>
Co-authored-by: Ronsard Carnegie <99965546+Carnegie2610@users.noreply.github.com>