feat(auth): role assignment and authorization (#20)
* feat(auth): role assignment and authorization

modify the prisma schema to keep the code consistant in the auth.middleware.ts and a generalize
getAuthUser file, making the schema uniform with the Ids, Email and user_role. Has modify all over
the code where the Id where involved, has created the TokenBlacklist model in the prisma for logging
out in the system. Test the role assignment in the teacher module and let the users with roles of
"admin" and 'teacher' only to be able to add new student in the system

* perf(module): teacher's authorization

modify the teacher's routes to let authorized users get access to specific endpoints, and also
modify the controller accordingly